Delete old code before refactoring
