+++ ui/src/components/RightMenu/components/right-menu-item.vue
