+++ ui/src/components/StreamEditor/index.vue
