+++ ui/src/components/StreamEditor/tagEditor.vue
