+++ ui/src/views/Stream/stream.vue
@@ -18,7 +18,7 @@
 -->
 
 <script setup>
-import Read from '@/components2/Read/index.vue'
+import Read from '@/components/Read/index.vue'
 </script>
 
 <template>
